requirements/local.txt
-r base.txt

django-debug-toolbar==1.2.2
ipdb==0.8
Werkzeug==0.9.6
